+++ src/sample_rate.rs
@@ -1,2 +1,2 @@
 // TODO remove this hard-coding
-pub(crate) const SAMPLE_RATE: f64 = 44100.;
+pub(crate) const SAMPLE_RATE: f32 = 44100.;
